# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nullryns (Øryns) API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  service?: string | null;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  service?: string | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface InquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @minLength 1 */
  serviceType: string;
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  timeline?: string | null;
  /** @minLength 1 */
  description: string;
}

export interface Inquiry {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  serviceType: string;
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  timeline?: string | null;
  description: string;
  status: string;
  isRead: boolean;
  createdAt: string;
}

export type InquiryStatusUpdateStatus = typeof InquiryStatusUpdateStatus[keyof typeof InquiryStatusUpdateStatus];


export const InquiryStatusUpdateStatus = {
  new: 'new',
  'in-discussion': 'in-discussion',
  quoted: 'quoted',
  closed: 'closed',
} as const;

export interface InquiryStatusUpdate {
  status: InquiryStatusUpdateStatus;
}

export interface ApiError {
  error: string;
}