# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nullryns (Øryns) API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: inquiries
    description: Project inquiry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submits a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List contact messages
      description: Returns all submitted contact messages
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
  /contact/{id}/read:
    patch:
      operationId: markContactMessageRead
      tags: [contact]
      summary: Mark contact message as read
      description: Marks a contact message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /inquiries/{id}/read:
    patch:
      operationId: markInquiryRead
      tags: [inquiries]
      summary: Mark inquiry as read
      description: Marks a project inquiry as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inquiry marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "404":
          description: Inquiry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /inquiries/{id}/status:
    patch:
      operationId: updateInquiryStatus
      tags: [inquiries]
      summary: Update inquiry status
      description: Updates the status of a project inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "404":
          description: Inquiry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /inquiries:
    post:
      operationId: submitInquiry
      tags: [inquiries]
      summary: Submit a project inquiry
      description: Submits a project brief / inquiry from a potential client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List project inquiries
      description: Returns all submitted project inquiries
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactInput:
      type: object
      required:
        - fullName
        - email
        - message
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        service:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1
    ContactMessage:
      type: object
      required:
        - id
        - fullName
        - email
        - message
        - isRead
        - createdAt
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        service:
          type: ["string", "null"]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InquiryInput:
      type: object
      required:
        - fullName
        - email
        - serviceType
        - description
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        serviceType:
          type: string
          minLength: 1
        budgetRange:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        description:
          type: string
          minLength: 1
    Inquiry:
      type: object
      required:
        - id
        - fullName
        - email
        - serviceType
        - description
        - status
        - isRead
        - createdAt
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        serviceType:
          type: string
        budgetRange:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        description:
          type: string
        status:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InquiryStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [new, in-discussion, quoted, closed]
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string