# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nullryns (Øryns) API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  service?: string | null;
  message: string;
  isRead: boolean;
  createdAt: Date;
}