# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nullryns (Øryns) API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  serviceType: string;
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  timeline?: string | null;
  description: string;
  status: string;
  isRead: boolean;
  createdAt: Date;
}